Better handling for duplicate replication docs.
Duplicate replication docs are documents which specify the same replication.

Handle 3 cases:

1) Documents are in same database, but potentially differen shards.

2) Documents are in two different replicator databases. For example in
   `a/_replicator` and `b/_replicator`. Handle this case better with a warning
   instead of crashing inside the case clause.

3) Handle the case when existing replication is transient, in other words,
    started from the `_replicate` endpoint.

A case that is not handled directly is replications which are from different
databases, different docs and different nodes. In that case they will collide
and contend on global registration of scheduler jobs.

Also fixed indentation in process_update.