Prevent ibrowse connections from being cleaned up too soon (#62)
After a request finishes, ibrowse will reset the state of the request and set
an idle timer. If connection is not reused in that time, it will be cleaned up
(process exits with `normal`).

In the scheduler pool we specify a custom idle connection timeout implied by
the `connection_close_interval` config setting, so make sure to set ibrowse's
idle timeout to match that, otherwise we are not getting the intended reuse
out of the pool.