Merge of "Inject random delays in scan_all_dbs" from master.
This functionality used to be in couch_replicator_manager but was moved to
couch_multidb_chanages module.

Add a random backoff, which get larger (but up to a minute) for each successive
database shard it finds. This is done in order to prevent a stampede during
startup. Especially if _replicator dbs need their VDUs updated.