Do not time-out scheduler's gen_server calls.
When there are a large number of jobs, reschedules can take longer than 5
seconds. So even though there is no socket or disk IO happening, because number
of jobs is unbounded do not rely on 5 seconds as a timeout.

Also, there is already a timeout handled as a temporary error by doc processor
worker. That one has a better error message and covers also other potentially
blocking cases such as fetching filter code from the source.