Improve speed scheduler's `add_job` API call
Make sure add_job returns quickly and doesn't block the scheduler.

Instead of first checking pending_jobs,  running a `select` query which is
which is O(n) in the size of the table, check first if table is already
larger than `max_jobs`.  If it is, skip the more expensive pending check.

This removes blockage and lets replicator's scheduler easily add up to 300k and
more per node.