Faster handling of crashed jobs in scheduler
When a job crashes, scheduler receives a `DOWN` message and tries to start
pending jobs. If there is a large number of total jobs in the ets table that
can take a significant amount of time. For example, even `pending_jobs_count/0`
is still an O(TotalJobs) operation. Combined with a cascading failure, where
lots of jobs in a short period of time, scheduler could be blocked for a very
long time.

To fix, as a heuristic, if total jobs is greater or equal to max jobs running,
assume table is large enough to avoid immediately starting pending jobs after
each crash, instead wait till next reschedule. (The same idea was used in
optimizing starting pending jobs when adding them, so use same strategy to be
consistent).