Remove transient jobs when they crash
Transient jobs are those started by a request to `_replicate` endpoint.

In older code those replication used to crash and were not restarted. So try to
restore the same behavior. This behavior is also consistent when nodes crash -
transient replications do not migrate to other nodes, while persistent ones do.