Handle duplicate jobs resulting from scheduler restarting
When scheduler is restarted by the supervisor but scheduler job sup is not, it
will try to re-add all the jobs and they will fails as duplicates. Handle this
case by accepting those jobs as running and recording their state as such.

Also maintain previous check for duplicate jobs running on another node as result
of duplicate global registrations. That case is handled as a crash still.