AIMD based rate limiter implementation
AIMD: additive increase / multiplicative decrease feedback control algorithm.

https://en.wikipedia.org/wiki/Additive_increase/multiplicative_decrease

This is an algorithm which converges on the available channel capacity.
Each participant doesn't a priori know the capacity and participants don't
communicate or know about each other (so they don't coordinate to divide
the capacity among themselves).

A variation of this is used in TCP congestion control algorithm. This is proven
to converge, while for example, additive increase / additive decrease  or
multiplicative increase / multiplicative decrease won't.

A few tweaks were applied to the base control logic:

 * Estimated value is an interval (period) instead of a rate. This is for
   convenience, as users will probably want to know how much to sleep. But,
   rate is just 1000 / interval, so it is easy to transform.

 * There is a hard max limit for estimated period.  Mainly as a practical concern
   as connections sleeping too long will timeout and / or jobs will waste time
   sleeping and consume scheduler slots, while others could be running.

 * There is a time decay component used to handle large pauses between updates.
   In case of large update interval, assume (optimistically) some successful
   requests have been made. Intuitively, the more time passes, the less accurate
   the estimated period probably is.

 * The rate of updates applied to the algorithm is limited. This effectively
   acts as a low pass filter and make the algorithm handle better spikes and
   short bursts of failures. This is not a novel idea, some alternative TCP
   control algorithms like Westwood+ do something similar.

 * There is a large downward pressure applied to the increasing interval as it
   approaches the max limit. This is done by tweaking the additive factor via
   a step function. In practice this has effect of trying to make it a bit
   harder for jobs to cross the maximum backoff threshold, as they would be
   killed and potentially lose intermediate work.

Main API functions are:

   success(Key) -> IntervalInMilliseconds

   failure(Key) -> IntervalInMilliseconds

   interval(Key) -> IntervalInMilliseconds

Key is any (hashable by phash2) term. Typically would be something like
{Method, Url}. The result from the function is the current period value. Caller
would then presumably choose to sleep for that amount of time before or after
making requests. The current interval can be read with interval(Key) function.

Implementation is sharded ETS tables based on the key and there is a periodic
timer which cleans unused items.