Add a job_summary/2 function to scheduler
`job_summary(JobId, HealthThreashold)` returns a short job summary suitable
for _scheduler/docs endpoint results. It doesn't contain the full history
details but instead has an `error_count` which is the number of consecutive
crashes a job has experienced.

`error_count` is an interesting value to show to the user as that is the
parameter scheduler will look at and decide if a job is crashing and needs to
be backed off (exponentially). Seeing a high error count will explain to the
user why their job hasn't been running for a long time even though there are
enough scheduling slots available.

The reason HealthThreshold is exposed via a separate function is because it
will avoid having to look up the config value for each job inside a fold
slowing down the request.