Improve replicator design doc update logic
Previously only VDU function was compared so that terminal_states was not
updated if VDU function didni't change. Even when it did change only the
VDU function code was updated but not the view.

Since replicator controls the whole _design/_replicator document, just compare
the whole document for change. This it is also more future proof. Any changes
will be updated.