Re-create missing design doc if it is not found.
Previously only created the document when a new replicator db was created or
was found. Howver, user can delete the design document anytime. When they,
query the _scheduler/docs endpoint, the query would crash as the view will
not be there.

So to fix, when querying the view and it is missing, try to re-create it.