Add start_time and last_updated fields to _scheduler/ endpoints output
Also for consistency add _replication_start_time field to completed and failed
replication document updates.

`start_time` is defined as the time since the replication first was noticed by
the replicator. For a document-based replication it is when the document update
event reached couch replicator and the replication record was parsed from it.
For a _replicate replication it is the time when the POST request body was
parsed.

`last_updated` is time of last state update.