Forward-port "Fix crashes when replicator db is deleted" from master
This is port of commit: 7a2b2b68bf4afa6d20b56a8ae51361f83981412a from master.

That commit was made to couch_replicator_manager so it had to be ported by
hand.

Most of the logic for replicator document handling is now in
couch_replicator_docs.erl module. The scanner part is in couch_multidb_changes.